server/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
boltons
Brotli
certifi
cffi
charset-normalizer
constantly==23.10.4
cryptography
cssselect==1.2.0
defusedxml==0.7.1
distro
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
filelock==3.15.4
frozendict
hyperlink==21.0.0
idna
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
jsonpatch
jsonpointer==2.1
lxml==5.3.0
numpy==2.1.0
packaging
pandas==2.2.2
parsel==1.9.1
platformdirs
pluggy
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycosat
pycparser
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
PySocks
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
queuelib==1.7.0
requests
requests-file==2.1.0
ruamel.yaml
Scrapy==2.11.2
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
tqdm
truststore
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3
w3lib==2.2.1
wheel==0.43.0
zope.interface==7.0.1
zstandard
